refactor(calendar): extract gantt bar constants and FloatingChip component
Move the flip-threshold and min-bar-width magic numbers into constants.ts,
and split FloatingChip out of ganttBar.tsx into its own file.
