+++ frontend/packages/app/src/pages/project-details/tabs/calendar/ganttBar.tsx
@@ -1,14 +1,14 @@
 /**
  * External dependencies.
  */
-import type { ComponentType, SVGProps } from "react";
 import { Tooltip } from "@rtcamp/frappe-ui-react";
 import { Sparkle, Zap } from "@rtcamp/frappe-ui-react/icons";
 
 /**
  * Internal dependencies.
  */
-import { COLUMN_WIDTH, MIN_CARD_DAYS } from "./constants";
+import { MIN_BAR_WIDTH } from "./constants";
+import { FloatingChip } from "./floatingChip";
 import type { ProjectTimelineItem } from "./types";
 
 type ItemPosition = { left: number; width: number };
@@ -18,60 +18,6 @@ type GanttBarProps = {
   pos: ItemPosition;
   totalWidth: number;
 };
-
-// Bars this narrow (a single-day span) can't fit their label inside, so the
-// label floats outside the icon chip instead of truncating.
-const MIN_BAR_WIDTH = COLUMN_WIDTH * MIN_CARD_DAYS;
-
-type FloatingChipProps = {
-  item: ProjectTimelineItem;
-  pos: ItemPosition;
-  totalWidth: number;
-  icon: ComponentType<SVGProps<SVGSVGElement>>;
-  chipClassName: string;
-  textClassName: string;
-};
-
-// Icon chip + floating label, flipped to the left when near the right edge.
-function FloatingChip({
-  item,
-  pos,
-  totalWidth,
-  icon: Icon,
-  chipClassName,
-  textClassName,
-}: FloatingChipProps) {
-  const spaceRight = totalWidth - (pos.left + pos.width);
-  const flipLeft = spaceRight < 150;
-
-  return (
-    <Tooltip text={item.title}>
-      <div
-        className="absolute top-1/2 -translate-y-1/2 z-1 flex items-center gap-2"
-        style={
-          flipLeft
-            ? {
-                right: totalWidth - (pos.left + pos.width),
-                flexDirection: "row-reverse",
-              }
-            : { left: pos.left }
-        }
-      >
-        <div
-          className={`flex items-center justify-center rounded-md mx-0.5 shrink-0 ${chipClassName}`}
-          style={{ width: pos.width, height: 32 }}
-        >
-          <Icon className="size-3.5" />
-        </div>
-        <span
-          className={`text-sm whitespace-nowrap ${textClassName}${item.isComplete ? " line-through opacity-60" : ""}`}
-        >
-          {item.title}
-        </span>
-      </div>
-    </Tooltip>
-  );
-}
 
 export function GanttBar({ item, pos, totalWidth }: GanttBarProps) {
   if (item.type === "Milestone") {
